Refactor: Navbar리팩토링
