+++ src/interfaces/menu.ts
@@ -1,0 +1,5 @@
+export interface Imenu {
+  id: number;
+  children: any;
+  route?: string;
+}
